+++ misskey-api/src/endpoint/antennas/delete.rs
@@ -20,25 +20,13 @@ mod tests {
 
     #[tokio::test]
     async fn request() {
-        use crate::model::{antenna::AntennaSource, query::Query};
-
         let client = TestClient::new();
         let antenna = client
-            .test(crate::endpoint::antennas::create::Request {
-                name: "test".to_string(),
-                src: AntennaSource::All,
-                user_list_id: None,
-                #[cfg(feature = "12-10-0")]
-                user_group_id: None,
-                keywords: Query::default(),
-                #[cfg(feature = "12-19-0")]
-                exclude_keywords: Query::default(),
-                users: Vec::new(),
-                case_sensitive: false,
-                with_replies: false,
-                with_file: false,
-                notify: false,
-            })
+            .test(
+                crate::endpoint::antennas::create::Request::builder()
+                    .name("test")
+                    .build(),
+            )
             .await;
 
         client