+++ misskey-api/src/streaming/channel/channel.rs
@@ -38,11 +38,11 @@ mod tests {
     async fn subscribe_unsubscribe() {
         let client = TestClient::new().await;
         let channel = client
-            .test(crate::endpoint::channels::create::Request {
-                name: "test".to_string(),
-                description: None,
-                banner_id: None,
-            })
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test")
+                    .build(),
+            )
             .await;
 
         let mut stream = client
@@ -58,11 +58,11 @@ mod tests {
     async fn stream() {
         let client = TestClient::new().await;
         let channel = client
-            .test(crate::endpoint::channels::create::Request {
-                name: "test".to_string(),
-                description: None,
-                banner_id: None,
-            })
+            .test(
+                crate::endpoint::channels::create::Request::builder()
+                    .name("test")
+                    .build(),
+            )
             .await;
 
         let mut stream = client
@@ -73,22 +73,12 @@ mod tests {
             .unwrap();
 
         future::join(
-            client.test(crate::endpoint::notes::create::Request {
-                visibility: None,
-                visible_user_ids: None,
-                text: Some("some text".to_string()),
-                cw: None,
-                via_mobile: None,
-                local_only: None,
-                no_extract_mentions: None,
-                no_extract_hashtags: None,
-                no_extract_emojis: None,
-                file_ids: None,
-                reply_id: None,
-                renote_id: None,
-                poll: None,
-                channel_id: Some(channel.id),
-            }),
+            client.test(
+                crate::endpoint::notes::create::Request::builder()
+                    .text("some text")
+                    .channel_id(channel.id)
+                    .build(),
+            ),
             async { stream.next().await.unwrap().unwrap() },
         )
         .await;