+++ misskey-util/CHANGELOG.md
@@ -11,12 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Registry APIs
 - Page APIs
+- Gallery APIs
+- Reactions APIs
 
 ### Changed
 ### Deprecated
 ### Removed
 
 - Latest version flag from being enabled as default
+- `ClientExt::users`
+   - For Misskey v12.88.0
+- `ClientExt::recommended_users`
+   - For Misskey v12.88.0 ~
+- `ClientExt::server_logs`
+   - For Misskey v12.93.0 ~
 
 ### Fixed
 ### Security